# webinterface/frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AllowedAction": {
        "enum": [
          "cancel",
          "retry_finalization",
          "replace_file",
          "delete_file",
          "launch",
          "relaunch_identical",
          "edit_and_relaunch",
          "delete_job",
          "regenerate_secret",
          "view_result"
        ],
        "title": "AllowedAction",
        "type": "string"
      },
      "ApiProblemCode": {
        "enum": [
          "bad_request",
          "forbidden",
          "resource_unavailable",
          "method_not_allowed",
          "conflict",
          "precondition_required",
          "rate_limited",
          "internal_error"
        ],
        "title": "ApiProblemCode",
        "type": "string"
      },
...
      "ErrorCode": {
        "enum": [
          "input_invalid",
          "input_too_large",
          "prompt_injection_detected",
          "prompt_security_check_failed",
          "transcription_failed",
          "processing_failed",
          "timeout",
          "cancel_failed",
          "server_interrupted",
          "artifact_write_failed",
          "result_integrity_failed"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiProblemCode"
              },
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ],
            "title": "Code"
          },
          "correlation_id": {
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Correlation Id",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PublicError"
            },
            "maxItems": 32,
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "maxLength": 256,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "maxLength": 256,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "correlation_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "PublicError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message_key": {
            "maxLength": 128,
            "pattern": "^[a-z0-9_.-]+$",
            "title": "Message Key",
            "type": "string"
          },
          "parameters": {
            "maxProperties": 16,
            "patternProperties": {
              "^[a-z0-9_-]+$": {
                "anyOf": [
                  {
                    "maxLength": 1024,
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  }
                ]
              }
            },
            "propertyNames": {
              "maxLength": 64
            },
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "code",
          "message_key"
        ],
        "title": "PublicError",
        "type": "object"
      },
...
      "SessionSnapshot": {
        "additionalProperties": false,
        "properties": {
          "allowed_actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AllowedAction"
            },
            "title": "Allowed Actions",
            "type": "array",
            "uniqueItems": true
          },
          "archive_excluded_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Archive Excluded Count",
            "type": "integer"
          },
          "archive_phase": {
            "anyOf": [
              {
                "enum": [
                  "transfer",
                  "extraction",
                  "track_validation",
                  "launch_preparation"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive Phase"
          },
          "context_text": {
            "default": "",
            "maxLength": 200000,
            "title": "Context Text",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicError"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UploadFileSnapshot"
            },
            "maxItems": 100,
            "title": "Files",
            "type": "array"
          },
          "input_type": {
            "default": "audio",
            "enum": [
              "audio",
              "merged_transcription",
              "zip"
            ],
            "title": "Input Type",
            "type": "string"
          },
          "language": {
            "pattern": "^[a-z]{2,3}(?:-[A-Z]{2})?$",
            "title": "Language",
            "type": "string"
          },
          "next_files_cursor": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Files Cursor"
          },
          "next_validations_cursor": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Validations Cursor"
          },
          "previous_summaries_text": {
            "default": "",
            "maxLength": 2000000,
            "title": "Previous Summaries Text",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "session_id": {
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UploadSessionStatus"
          },
          "validations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationSnapshot"
            },
            "maxItems": 100,
            "title": "Validations",
            "type": "array"
          }
        },
        "required": [
          "session_id",
          "revision",
          "status",
          "expires_at",
          "language"
        ],
        "title": "SessionSnapshot",
        "type": "object"
      },
...
      "UploadSessionStatus": {
        "enum": [
          "created",
          "uploading",
          "validating",
          "ready",
          "waiting_for_capacity",
          "consumed",
          "cancelled",
          "expired"
        ],
        "title": "UploadSessionStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Tara Web API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tara-Job-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tara-Job-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/v1/uploads/sessions": {
      "post": {
        "operationId": "create_session_api_v1_uploads_sessions_post",
        "parameters": [
          {
            "in": "query",
            "name": "input_type",
            "required": false,
            "schema": {
              "default": "audio",
              "enum": [
                "audio",
                "merged_transcription",
                "zip"
              ],
              "title": "Input Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tara-creation-recovery",
            "required": true,
            "schema": {
              "title": "X-Tara-Creation-Recovery",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uniform protected API error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}